refactor metadata_formats tests to allow growth
when adding a metadata formatter in the future, make a test class
extending BaseMetadataFormatterTest so all our formatters can share a
set of test cases
